Refactor salário mínimo
