+++ cacimbao/__init__.py
@@ -1,4 +1,4 @@
-from cacimbao.datasets.loaders import download_dataset, load_dataset
-from cacimbao.new_datasets import list_datasets
+from cacimbao.datasets import list_datasets
+from cacimbao.loaders import download_dataset, load_dataset
 
 __all__ = ["download_dataset", "list_datasets", "load_dataset"]
